requirements.txt
requests==2.31.0
twilio
openai
flask
python-dotenv
